requirements.txt
accelerate
langchain
langchain-groq
python-dotenv
streamlit
